+++ event-sourcing/event-publisher/event-publisher-timer/src/main/java/uk/gov/justice/services/eventsourcing/publishedevent/prepublish/PrePublishTimerBean.java
@@ -1,6 +1,6 @@
 package uk.gov.justice.services.eventsourcing.publishedevent.prepublish;
 
-import uk.gov.justice.services.eventsourcing.util.jee.timer.TimerCanceler;
+import uk.gov.justice.services.eventsourcing.util.jee.timer.StopWatchFactory;
 import uk.gov.justice.services.eventsourcing.util.jee.timer.TimerServiceManager;
 
 import javax.annotation.PostConstruct;
@@ -11,31 +11,32 @@ import javax.ejb.Timeout;
 import javax.ejb.TimerService;
 import javax.inject.Inject;
 
+import org.apache.commons.lang3.time.StopWatch;
+
 @Singleton
 @Startup
 public class PrePublishTimerBean {
 
-    private static final int THRESHOLD = 10;
     private static final String TIMER_JOB_NAME = "event-store.pre-publish-events.job";
 
     @Resource
-    TimerService timerService;
+    private TimerService timerService;
 
     @Inject
-    PrePublishTimerConfig prePublishTimerConfig;
+    private PrePublishTimerConfig prePublishTimerConfig;
 
     @Inject
-    TimerServiceManager timerServiceManager;
+    private TimerServiceManager timerServiceManager;
 
     @Inject
-    TimerCanceler timerCanceler;
+    private PrePublishProcessor prePublishProcessor;
 
     @Inject
-    PrePublishProcessor prePublishProcessor;
+    private StopWatchFactory stopWatchFactory;
 
     @PostConstruct
     public void startTimerService() {
-        timerCanceler.cancelTimer(TIMER_JOB_NAME, timerService);
+
         timerServiceManager.createIntervalTimer(
                 TIMER_JOB_NAME,
                 prePublishTimerConfig.getTimerStartWaitMilliseconds(),
@@ -46,15 +47,14 @@ public class PrePublishTimerBean {
     @Timeout
     public void performPrePublish() {
 
-        final int maxEventsPublishedPerIteration = prePublishTimerConfig.getMaxEventsPublishedPerIteration();
+        final long maxRuntimeMilliseconds = prePublishTimerConfig.getTimerMaxRuntimeMilliseconds();
+        final StopWatch stopWatch = stopWatchFactory.createStopWatch();
 
-        int numberOfEventsPublished = 1;
+        stopWatch.start();
+
         while (prePublishProcessor.prePublishNextEvent()) {
 
-            timerServiceManager.cancelOverlappingTimers(TIMER_JOB_NAME, THRESHOLD, timerService);
-            numberOfEventsPublished++;
-
-            if (numberOfEventsPublished > maxEventsPublishedPerIteration) {
+            if (stopWatch.getTime() > maxRuntimeMilliseconds) {
                 break;
             }
         }