+++ event-sourcing/event-publisher/event-publisher-timer/src/test/java/uk/gov/justice/services/eventsourcing/publishedevent/prepublish/PrePublishTimerConfigTest.java
@@ -36,12 +36,12 @@ public class PrePublishTimerConfigTest {
     }
 
     @Test
-    public void shouldGetTheMaxNumberOfEventsPublishedPerIteration() throws Exception {
+    public void shouldGetTheTimerMaxRuntime() throws Exception {
 
-        final int maxEventsPublishedPerIteration = 23;
+        final long milliseconds = 600L;
 
-        setField(prePublishTimerConfig, "maxEventsPublishedPerIteration", "" + maxEventsPublishedPerIteration);
+        setField(prePublishTimerConfig, "timerMaxRuntimeMilliseconds", "" + milliseconds);
 
-        assertThat(prePublishTimerConfig.getMaxEventsPublishedPerIteration(), is(maxEventsPublishedPerIteration));
+        assertThat(prePublishTimerConfig.getTimerMaxRuntimeMilliseconds(), is(milliseconds));
     }
 }