+++ event-sourcing/event-publisher/event-publisher-timer/src/test/java/uk/gov/justice/services/eventsourcing/publishedevent/publishing/PublisherTimerBeanTest.java
@@ -1,15 +1,17 @@
 package uk.gov.justice.services.eventsourcing.publishedevent.publishing;
 
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import uk.gov.justice.services.eventsourcing.publishedevent.publish.PublishedEventDeQueuerAndPublisher;
-import uk.gov.justice.services.eventsourcing.util.jee.timer.TimerCanceler;
+import uk.gov.justice.services.eventsourcing.util.jee.timer.StopWatchFactory;
 import uk.gov.justice.services.eventsourcing.util.jee.timer.TimerServiceManager;
 
 import javax.ejb.TimerService;
 
+import org.apache.commons.lang3.time.StopWatch;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
@@ -32,7 +34,7 @@ public class PublisherTimerBeanTest {
     private TimerServiceManager timerServiceManager;
 
     @Mock
-    private TimerCanceler timerCanceler;
+    private StopWatchFactory stopWatchFactory;
 
     @InjectMocks
     private PublisherTimerBean publisherTimerBean;
@@ -40,15 +42,14 @@ public class PublisherTimerBeanTest {
     @Test
     public void shouldSetUpTheTimerServiceOnPostConstruct() throws Exception {
 
-        final long timerStartValue = 7000;
-        final long timerIntervalValue = 2000;
+        final long timerStartValue = 7000L;
+        final long timerIntervalValue = 2000L;
 
         when(publisherTimerConfig.getTimerStartWaitMilliseconds()).thenReturn(timerStartValue);
         when(publisherTimerConfig.getTimerIntervalMilliseconds()).thenReturn(timerIntervalValue);
 
         publisherTimerBean.startTimerService();
 
-        verify(timerCanceler).cancelTimer("event-store.de-queue-events-and-publish.job", timerService);
         verify(timerServiceManager).createIntervalTimer(
                 "event-store.de-queue-events-and-publish.job",
                 timerStartValue,
@@ -59,11 +60,34 @@ public class PublisherTimerBeanTest {
     @Test
     public void shouldRunPublishUntilAllEventsArePublished() throws Exception {
 
+        final long timerIntervalValue = 2000L;
+        final long timerMaxRuntimeValue = 495L;
+
+        when(publisherTimerConfig.getTimerIntervalMilliseconds()).thenReturn(timerIntervalValue);
+        when(publisherTimerConfig.getTimerMaxRuntimeMilliseconds()).thenReturn(timerMaxRuntimeValue);
+        when(stopWatchFactory.createStopWatch()).thenReturn(mock(StopWatch.class));
         when(publishedEventDeQueuerAndPublisher.deQueueAndPublish()).thenReturn(true, true, false);
 
         publisherTimerBean.doDeQueueAndPublish();
 
         verify(publishedEventDeQueuerAndPublisher, times(3)).deQueueAndPublish();
-        verify(timerServiceManager, times(2)).cancelOverlappingTimers("event-store.de-queue-events-and-publish.job", 10, timerService);
+    }
+
+    @Test
+    public void shouldRunPublishUntilTimeRunsOut() throws Exception {
+
+        final long timerIntervalValue = 2000L;
+        final long timerMaxRuntimeValue = 459L;
+        final StopWatch stopWatch = mock(StopWatch.class);
+
+        when(publisherTimerConfig.getTimerIntervalMilliseconds()).thenReturn(timerIntervalValue);
+        when(publisherTimerConfig.getTimerMaxRuntimeMilliseconds()).thenReturn(timerMaxRuntimeValue);
+        when(stopWatchFactory.createStopWatch()).thenReturn(stopWatch);
+        when(publishedEventDeQueuerAndPublisher.deQueueAndPublish()).thenReturn(true, true);
+        when(stopWatch.getTime()).thenReturn(timerIntervalValue);
+
+        publisherTimerBean.doDeQueueAndPublish();
+
+        verify(publishedEventDeQueuerAndPublisher, times(1)).deQueueAndPublish();
     }
 }